// src/app/modules/hcm/shared/models/laborRelationship/labor-relationship-grouping.ts
export class LaborRelationshipxGrouping{
    
    idLaborRelationshipxGrouping: number = -1;  
    idGrouping: number = -1;    
    idLaborRelationship: number = -1;    
    createdByUserId: number = 0;    
    updatedByUserId: number = 0;    
    assignedValue: string = "";
    groups: string = "";
    abbreviation: string = "";
    active: boolean = true;
}